pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Yutarotaro"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.1"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
python-dotenv = "^0.19.0"
SQLAlchemy = "^1.4.25"
aiomysql = "^0.0.21"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
